refactor: :recycle: account, token 폴더 구조 변경
- #210
