+++ app/src/login/login.service.ts
@@ -13,9 +13,9 @@ import { Cron } from '@nestjs/schedule';
 import { OAuth2Client } from 'google-auth-library';
 import mongoose from 'mongoose';
 import { lastValueFrom } from 'rxjs';
-import { AccountService } from 'src/api/account/account.service';
-import type { token } from 'src/api/token/db/token.database.schema';
-import { TokenService } from 'src/api/token/token.service';
+import { AccountService } from 'src/login/account/account.service';
+import type { token } from 'src/auth/token/db/token.database.schema';
+import { TokenService } from 'src/auth/token/token.service';
 import type { FtClientConfig } from 'src/config/configuration/ftClient.config';
 import type { GoogleClientConfig } from 'src/config/configuration/googleClient.config';
 import type { JwtConfig } from 'src/config/configuration/jwt.config';
